refactor: simplify layout
