+++ src/components/pages/home/index.js
@@ -1,7 +1,6 @@
 import Hero from './hero'
 import Screenshots from './screenshots'
 import Meta from './meta'
-import Explore from './explore'
 import Analytics from './analytics'
 
-export { Hero, Screenshots, Meta, Explore, Analytics }
+export { Hero, Screenshots, Meta, Analytics }
